Add proper snapper root set up.
Assuming a root subvolume of `@` (configurable), we set up snapper to
allow automatic snapshots in case a system update breaks our machine.

This is extremely important if I want to use Arch on a work machine.

Also provide documentation of how it all fits together, when I
inevitably try to tweak this later.