Remove insertion of lines into mkinitcpio.conf
This ruined our idempotency, because every run we'd replace
/etc/mkinitcpio.conf with a blank copy and then amend it.

If we're serious about making this "pluggable", we likely need to set
the modules/hooks as facts and dynamically load in other tasks to tweak
this accordingly.

But the good news is that this doesn't seem necessary, so we can just
plop in a stock `mkinitcpio` and go about our day.